Major project refactor
