+++ src/assets/sass/base/_fonts.scss
@@ -1,0 +1,48 @@
+@font-face {
+    font-family: 'Chloe-Regular';
+    src: url('../../fonts/Chloe/Regular/Chloe-Regular.eot'); /* IE9 Compat Modes */
+    src: url('../../fonts/Chloe/Regular/Chloe-Regular.eot?#iefix') format('embedded-opentype'), /* IE6-IE8 */
+         url('../../fonts/Chloe/Regular/Chloe-Regular.woff') format('woff'), /* Modern Browsers */
+         url('../../fonts/Chloe/Regular/Chloe-Regular.ttf')  format('truetype'), /* Safari, Android, iOS */
+         url('../../fonts/Chloe/Regular/Chloe-Regular.svg#svgChloe-Regular') format('svg'); /* Legacy iOS */
+}
+@font-face {
+    font-family: 'Adamina-Regular';
+    src: url('../../fonts/Adamina/Regular/Adamina-Regular.eot'); /* IE9 Compat Modes */
+    src: url('../../fonts/Adamina/Regular/Adamina-Regular.eot?#iefix') format('embedded-opentype'), /* IE6-IE8 */
+         url('../../fonts/Adamina/Regular/Adamina-Regular.woff') format('woff'), /* Modern Browsers */
+         url('../../fonts/Adamina/Regular/Adamina-Regular.ttf')  format('truetype'), /* Safari, Android, iOS */
+         url('../../fonts/Adamina/Regular/Adamina-Regular.svg#svgAdamina-Regular') format('svg'); /* Legacy iOS */
+}
+@font-face {
+    font-family: 'ProximaNova-Thin';
+    src: url('../../fonts/ProximaNova/Thin/ProximaNova-Thin.eot'); /* IE9 Compat Modes */
+    src: url('../../fonts/ProximaNova/Thin/ProximaNova-Thin.eot?#iefix') format('embedded-opentype'), /* IE6-IE8 */
+         url('../../fonts/ProximaNova/Thin/ProximaNova-Thin.woff') format('woff'), /* Modern Browsers */
+         url('../../fonts/ProximaNova/Thin/ProximaNova-Thin.ttf')  format('truetype'), /* Safari, Android, iOS */
+         url('../../fonts/ProximaNova/Thin/ProximaNova-Thin.svg#svgProximaNova-Thin') format('svg'); /* Legacy iOS */
+}
+@font-face {
+    font-family: 'ProximaNova-Regular';
+    src: url('../../fonts/ProximaNova/Regular/ProximaNova-Regular.eot'); /* IE9 Compat Modes */
+    src: url('../../fonts/ProximaNova/Regular/ProximaNova-Regular.eot?#iefix') format('embedded-opentype'), /* IE6-IE8 */
+         url('../../fonts/ProximaNova/Regular/ProximaNova-Regular.woff') format('woff'), /* Modern Browsers */
+         url('../../fonts/ProximaNova/Regular/ProximaNova-Regular.ttf')  format('truetype'), /* Safari, Android, iOS */
+         url('../../fonts/ProximaNova/Regular/ProximaNova-Regular.svg#svgProximaNova-Regular') format('svg'); /* Legacy iOS */
+}
+@font-face {
+    font-family: 'ProximaNova-Semibold';
+    src: url('../../fonts/ProximaNova/SemiBold/ProximaNova-Semibold.eot'); /* IE9 Compat Modes */
+    src: url('../../fonts/ProximaNova/SemiBold/ProximaNova-Semibold.eot?#iefix') format('embedded-opentype'), /* IE6-IE8 */
+         url('../../fonts/ProximaNova/SemiBold/ProximaNova-Semibold.woff') format('woff'), /* Modern Browsers */
+         url('../../fonts/ProximaNova/SemiBold/ProximaNova-Semibold.ttf')  format('truetype'), /* Safari, Android, iOS */
+         url('../../fonts/ProximaNova/SemiBold/ProximaNova-Semibold.svg#svgProximaNova-SemiBold') format('svg'); /* Legacy iOS */
+}
+@font-face {
+    font-family: 'ProximaNova-Bold';
+    src: url('../../fonts/ProximaNova/Bold/ProximaNova-Bold.eot'); /* IE9 Compat Modes */
+    src: url('../../fonts/ProximaNova/Bold/ProximaNova-Bold.eot?#iefix') format('embedded-opentype'), /* IE6-IE8 */
+         url('../../fonts/ProximaNova/Bold/ProximaNova-Bold.woff') format('woff'), /* Modern Browsers */
+         url('../../fonts/ProximaNova/Bold/ProximaNova-Bold.ttf')  format('truetype'), /* Safari, Android, iOS */
+         url('../../fonts/ProximaNova/Bold/ProximaNova-Bold.svg#svgProximaNova-Bold') format('svg'); /* Legacy iOS */
+}
