+++ src/assets/sass/components/_carousel.scss
