+++ src/assets/sass/components/_newsletter.scss
