+++ src/assets/sass/main.scss
@@ -1,0 +1,20 @@
+//Base Import
+@import 'base/reset';
+@import 'base/fonts';
+
+//Components Import
+@import 'components/navbar';
+@import 'components/buttons';
+@import 'components/audio-player';
+@import 'components/carousel';
+@import 'components/newsletter';
+@import 'components/cards';
+
+//Pages Import
+@import 'views/page';
+
+//Themes Import
+@import 'themes/theme';
+
+//Layout Import
+@import 'layout/mods';
