+++ src/assets/sass/themes/_theme.scss
@@ -1,0 +1,3 @@
+body{
+    background-color: #FFFFFF;
+}
