+++ src/assets/sass/views/_page.scss
@@ -1,6 +1,3 @@
-//Mods
-@import './mods';
-
 .page{
     display: flex;
     align-items: center;
