+++ src/components/AudioPlayer/index.jsx
@@ -3,9 +3,6 @@ import React, { useState,useEffect } from 'react';
 //Components
 import VolumePlayerSlider from './VolumePlayerSlider';
 
-//Styles
-import './_audio-player.scss';
-
 const AudioPlayer = (props) => {
     const [audio] = useState(new Audio(props.audioURL));
     const [isAudioPlaying, setIsAudioPlaying] = useState(false);
