+++ src/components/Carousel/index.jsx
@@ -4,9 +4,6 @@ import React, {useState} from 'react';
 import CarouselController from './CarouselController';
 import CarouselRadioList from './CarouselRadioList';
 
-//Styles
-import './_carousel.scss';
-
 const Carousel = (props) => {
     const [currentPage,setCurrentPage] = useState(0);
     //When there's only one children props.children return the children not an array with that children.
